SBS initial commit
Introducing the simple build system (SBS). Designed to be an easy-to-use
C/C++ build system that spares codes the tedious details of Makefiles,
and concentrate on the actual settings of the build:

1. What the system needs to build
2. What settings the build should use
3. What order the build will occur

The system assumes only 2 prerequisites: gcc (or g++) and Gnu-make.

Detailing all the system aspects in a commit message will be somewhat
time consuming and probably not useful, so just a few major design
points:

1. The entire system is actually one file - the 'module.inc.sh' file. In
   order for one to use SBS all she needs to do is to put the file in
   her project root directory. Beyond that point its a matter as minimal
   as possible definitions within Makefiles.
2. SBS is designed with simplicity in mind. A minimal SBS Makefile is
   consisted of 2 lines:

   	MODULE_OBJS := main.o
   	include module.inc.mk

   Such a Makefile will try to create an executable from a source file
   name 'main.c', using gcc, and will be named after the directory the
   file resides on. Almost very part of this default configuration can
   be modified.
3. SBS allows easy modification of the GCC's flags, preprocessor
   definitions, warnings and linker flags
4. SBS multi directories projects, with the ability to:
	- Build directories in predefined order
	- Define a project level definitions file to provide common
	  settings to all directories
*. And more

Further details of exactly how to use systems will be provided in
future documentation and readme patches.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>