Modules additional dependencies setting
Adds an option to a module's Makefile to define a variable names
MODULE_BIN_DEPS. If such a variable is set, linkage of the binary is
decedent on it as well. By itself, the variable *doesn't* guarantee the
dependency existence, and it is up to the coder to make sure it exist in
order for the build to complete.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>