Improve binaries test structure
Have a lib/<FLAVOR> libraries instead of just lib. Making this closer to
a real life project where one need to distinguish between different
artifacts of different flavors.

No changes to the SBS itself, just some small adjustments to the
libraries locations and definitions of the build so library path is at
'lib/dbg' instead of just 'lib' (the binaries build is defaulted to dbg
flavor).

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>