Module binary dependencies
Add setting for binary dependencies that are needs to be linked with the
target binary (gcc setting), but also serve as it's build dependency
(make setting).

This sort of setting is useful if when linking a binary with a library
or an object that keeps changing, maybe in the same project but in a
different module (directory).

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>