sbs: use source list instead of object list
Replace the previous object lists as the base of the source lists to
compile with s source lists that explicitly set which source to build.
This is easier to read and understand, makes the entire languages mixing
(i.e. C and C++) much easier, and simplifies the entire source suffixes
settings.

Note that a file named file.c is now compiled to file.c.o instead of
just file.o, so one can have multiple file.<SUFFIX> in the same folder
(thought it's probably not a very goo idea in the first place).

Also - this patch kills the confusing MODULE_LANG setting. It is no
longer needed. Files are compiled according to their file type which is
determined by their suffix.

Linkage is of a module's artifact is usually done by a linker tool that
is deducted by the SBS: if any C++ source files are involved, g++ is
used. Otherwise, gcc is used. This can be overriden by the newly added
MODULE_LD setting for the loader, though in most cases using this
setting isn't required and should be avoided.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>