test: multi-lang2 test
A slightly more complicated test with both C and C++ source files. The
linker (MODULE_LD), is chosen automatically by SBS. Note that none-C++
linker (i.e. gcc instead of g++) will create a linking error, as the it
(the linker) will not find libstdc++ library symbols to link with.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>