sbs: allow external objects for an artifact
Allow modules to define external objects to be dependent on. External
objects are not being built by the module, and expected to exist prior
to the module build. They are also used as dependencies to the module
binary so a link step is taken if the and external object is updated.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>