sbs: rename ALL_TARGET to MAIN_TARGET
This is a leftover from the beginning, and now a confusing one, as
there's an 'all' target that is the default target, but it is not the
one refer to by ALL_TARGET, which is either the artifact target, or an
empty one if no artifact exists. So 'MAIN_TARGET' seems much more
appropriate.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>