sbs: sub module dependencies graph rehaul
Instead of the previous naive approach to sub modules building, SBS now
further uses Makes dependencies graph ability to build all the sub
modules, hence instead of a recipe that invokes multiple make
invocation, required parts of the build are built in the desired order
by using rules dependencies. This makes the entire build process much
faster for two reasons:

1. Internal - Many recursive make calls are being avoided, and are
   handled internally by SBS.
2. External - All pre and post sub modules can now be built together
   instead of one by one, which can impact the speed significantly in
   projects that have many of these sub-modules.

As a brief benchmark, running the test suite of SBS takes about 50% of
the time it used to before this change.

This change has quite a lot of implications on how the build is
executed. In particular, pre and post modules are no longer required to
be build 1 by 1, but can be built in parallel as well.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>