test: fix binary with external objects test
Since the binary depends on another sum module (nobin) to be built
before itself, some settings were missing. As the comment in the test
itself explains, the external objects feature doesn't do by default
anything regarding the objects themselves. It's the developer
responsibility to make sure they exist

In our case, we require the pre sub module to be built, however in this
case it isn't enough, as the objects appear in the module's artifact
dependency list, and make will fail as it doesn't know what to do with
their absence, hence the additional dependency we added between the
objects and the module that will build them, enforcing the right order
of objects generation.

This is the 'right' solution for SBS based external objects (a case that
doesn't make much sense to begin with). User specific cases will require
different dependencies to make sure the order is maintained.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>